# backend/docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Backend API
  description: API documentation for the backend service
  version: "1.0.0"
servers:
  - url: http://localhost:8888
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
...
    ResponseObject:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Success
        data:
          nullable: true
    ResponsePage:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    Pagination:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
...
    ProductRequest:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        brandId:
          type: integer
        categoryId:
          type: integer
        description:
          type: string
        unit:
          type: string
        trackingType:
          type: string
        sellPrice:
          type: number
          format: decimal
        minStock:
          type: integer
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
          nullable: true
        barcode:
          type: string
          nullable: true
        brandId:
          type: integer
          nullable: true
        brandName:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        trackingType:
          type: string
          nullable: true
        sellPrice:
          type: number
          format: decimal
          nullable: true
        minStock:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # =========================== SUPPLIER ===========================
    SupplierRequest:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        taxCode:
          type: string
        note:
          type: string
    SupplierResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        taxCode:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # =========================== PRODUCT IMAGE ===========================
...
    ProductImageResponse:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        url:
          type: string
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time

    # =========================== AUDIT LOG ===========================
...
    CustomerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        note:
          type: string
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # =========================== LOCATION ===========================
...
  parameters:
    pageParam:
      name: page
      in: query
      description: Page number (0-indexed)
      schema:
        type: integer
        default: 0
    sizeParam:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        default: 20
    sortParam:
      name: sort
      in: query
      description: Sort field (e.g. `name,asc`)
      schema:
        type: string
paths:
...
  /api/v1/product:
    get:
      tags: [Product]
      summary: List products (paginated)
      operationId: listProducts
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
        - $ref: "#/components/parameters/sortParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
    post:
      tags: [Product]
      summary: Create a product
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ProductResponse"
  /api/v1/product/{id}:
    get:
      tags: [Product]
      summary: Get product detail
      operationId: getProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ProductResponse"
    put:
      tags: [Product]
      summary: Update a product
      operationId: updateProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ProductResponse"
  /api/v1/product/{id}/toggle-active:
    put:
      tags: [Product]
      summary: Toggle product active status
      operationId: toggleProductActive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ProductResponse"

  # =========================== SUPPLIER ===========================
  /api/v1/supplier:
    get:
      tags: [Supplier]
      summary: List suppliers (paginated)
      operationId: listSuppliers
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
        - $ref: "#/components/parameters/sortParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
    post:
      tags: [Supplier]
      summary: Create a supplier
      operationId: createSupplier
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/SupplierResponse"
  /api/v1/supplier/{id}:
    get:
      tags: [Supplier]
      summary: Get supplier detail
      operationId: getSupplier
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/SupplierResponse"
    put:
      tags: [Supplier]
      summary: Update a supplier
      operationId: updateSupplier
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/SupplierResponse"
  /api/v1/supplier/{id}/toggle-active:
    put:
      tags: [Supplier]
      summary: Toggle supplier active status
      operationId: toggleSupplierActive
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/SupplierResponse"

  # =========================== PRODUCT IMAGE ===========================
...
  /api/v1/product-image/product/{productId}:
    get:
      tags: [Product Image]
      summary: List images for a product
      operationId: getProductImages
      security:
        - BearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProductImageResponse"
    delete:
      tags: [Product Image]
      summary: Delete all images for a product
      operationId: deleteProductImagesByProductId
      security:
        - BearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseObject"
...
  /api/v1/product-unit/product/{productId}:
    get:
      tags: [Product Unit]
      summary: List product units by product
      operationId: getProductUnitsByProduct
      security:
        - BearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"

  # =========================== STOCK CHECK ===========================
...
  /api/v1/customer:
    post:
      tags: [Customer]
      summary: Create a new customer
      operationId: createCustomer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/CustomerResponse"

    get:
      tags: [Customer]
      summary: List customers (paginated)
      operationId: listCustomers
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
  /api/v1/customer/search:
    get:
      tags: [Customer]
      summary: Search customers by keyword
      operationId: searchCustomers
      security:
        - BearerAuth: []
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"
  /api/v1/customer/{id}:
    get:
      tags: [Customer]
      summary: Get customer by ID
      operationId: getCustomer
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/CustomerResponse"

    put:
      tags: [Customer]
      summary: Update customer
      operationId: updateCustomer
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseObject"

    delete:
      tags: [Customer]
      summary: Delete customer (soft)
      operationId: deleteCustomer
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseObject"

  # =========================== LOCATION ===========================
...
  /api/v1/inventory:
    get:
      tags: [Inventory]
      summary: List inventory items (paginated)
      operationId: listInventoryItems
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseObject"
                  - properties:
                      data:
                        $ref: "#/components/schemas/ResponsePage"

  # =========================== REPORT ===========================
...